New grass, globals, menu handling
Created a new grass image and removed the old one.
Added width and height globals (testbed.WIDTH, testbed.HEIGHT)
Changed the Menu in a way, that it takes an array of items with scenes
Fixed world sprites so they have their anchor upper left which makes
more sense for the way I generate the world view.