Breaked the working!
These changes do not do what they're supposed to.
I want to have the Size of each world tile fixed and independed of the
number of world tiles. Up to now they were dependend.
Now they are not but still not the size I want them to have... or they
are and I miss something :)